fix(page): wire official api overrides through page commands
