+++ cmd/root.go
@@ -1,5 +1,7 @@
 package cmd
 
+import "github.com/lox/notion-cli/internal/config"
+
 type Context struct {
 	JSON             bool
 	Token            string
@@ -30,4 +32,15 @@ type VersionCmd struct {
 func (c *VersionCmd) Run(ctx *Context) error {
 	println("notion-cli version " + c.Version)
 	return nil
+}
+
+func officialAPIOverrides(ctx *Context) config.APIOverrides {
+	if ctx == nil {
+		return config.APIOverrides{}
+	}
+	return config.APIOverrides{
+		BaseURL:       ctx.APIBaseURL,
+		NotionVersion: ctx.APINotionVersion,
+		Token:         ctx.APIToken,
+	}
 }
